+++ main/CMakeLists.txt
@@ -2,15 +2,17 @@ idf_component_register(
   SRCS
     "main.c"
     "keypad.c"
-    "panel-attest.c"
+    "panel-connect.c"
+    "panel-game.c"
     "panel-gifs.c"
+    "panel-keyboard.c"
     "panel-menu.c"
     "panel-space.c"
+    "pixels.c"
     "task-ble.c"
+    "utils.c"
     "system/device-info.c"
-    "system/pixels.c"
-    "system/utils.c"
-    "utils/cbor.c"
+
   INCLUDE_DIRS
     ""
 )
